view: SAPPHIRE mask phasing bit on first allele
XCFTools does add a phasing bit to the first allele, this is not
standard under the BCF binary representation of VCF. This is
documented in section 6.3.3 of The Variant Call Format Specification
VCFv4.3 and BCFv2.2 from 27 Nov 2022 [1]. The first allele does not
carry the phase bit, for example VCF "0|1" is encoded as "0x02 05"
in BCF binary format. Some software, e.g., XCFTools can add the
phase bit to the first allele, therefore encoding "0|1" as "0x03 05"

This is usually not an issue as this bit does not carry any relevant
information in BCF, both "0x02 05" and "0x03 05" will be converted
to "0|1" in VCF as the first bit is discarded.

However, this bit gets copied into the SAPPHIRE binary file and
therefore extraction from VCF, BCF, and XCF might generated files
that differ by this bit, this makes it impossible to quickly
compare the binary files for equality through a checksum, therefore
SAPPHIRE binary extraction masks this bit, so that the binary files
extracted from VCF, BCF, and XCF are exactly the same.

Signed-off-by: Rick Wertenbroek <rick.wertenbroek@gmail.com>

[1] https://samtools.github.io/hts-specs/VCFv4.3.pdf